refactor(config): move PochiConfigFile implementation to a separate file and remove obsolete config schema
